polling: Introduce API for exclusive thread wake up
The poll infrastructure has no way to signal that a particular
file descriptor with ready events should only wake a single
thread.  It is currently not defined whether such a capability
is possible.  One issue that can occur is the Thundering Herd
issue, which will wake each thread for even a single ready
event.

This change introduces an API to indicate that the poll block
infrastructure should honor a wait to only wake a single thread
if possible.  The change here is API only - no change to the
actual behavior is implemented at the moment.  There are no
users of the new API introduced with this patch.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Cc: David Ahern <dahern@digitalocean.com>
Cc: Matteo Croce <technoboy85@gmail.com>
Cc: Anton Ivanov <anton.ivanov@cambridgegreys.com>